Share the parts both feature file tools agree on
`extract-feature-php.php` and `phpstan-feature-files.php` each carried
their own copy of the checks deciding which docstrings hold a PHP block and
where a block may be extracted to, in the same namespace and byte for byte
the same. Nothing catches such a pair drifting apart: the two run as
separate processes, `utils` is not autoloaded, and the static analysis only
looks at `src` and `tests`. The copies had already started to diverge, with
a fix for an extraction target resolving to the filesystem root landing in
one of them first.

They move to `utils/feature-php-blocks.php`, which both scripts pull in
themselves. It stays out of the autoloader, as loading a set of functions
into every package using the framework buys nothing.

`remove_extracted_files()` takes the pattern matching the extracted file
names as an argument, since the two tools name their files differently, and
the manifest that only the analysis writes is removed by the analysis.
